refactor(wallets): create extendable AVMProvider for Kibisis client
This refactors the Kibisis wallet client to use an extendable `AVMProvider` class, laying the groundwork for future wallet implementations that also use `@agoralabs-sh/avm-web-provider`.

- Created new `AVMProvider` abstract class
- Refactored `KibisisWallet` to extend `AVMProvider`
- Moved common AVM Web Provider functionality to `AVMProvider`
- Implemented abstract methods in `KibisisWallet` for Kibisis-specific behavior
